Move SchemaComparer and FieldComparer to their own files
