+++ csharp/test/Apache.Arrow.Tests/SchemaBuilderTests.cs
@@ -23,34 +23,6 @@ namespace Apache.Arrow.Tests
 {
     public class SchemaBuilderTests
     {
-        public class SchemaComparer
-        {
-            public static bool Equals(Schema s1, Schema s2)
-            {
-                if (ReferenceEquals(s1, s2))
-                {
-                    return true;
-                }
-                if (s2 == null || s1 == null || s1.HasMetadata != s2.HasMetadata || s1.Fields.Count != s2.Fields.Count)
-                {
-                    return false;
-                }
-
-                if (!s1.Fields.Keys.All(k => s2.Fields.ContainsKey(k) && FieldComparer.Equals(s1.Fields[k], s2.Fields[k])) ||
-                    !s2.Fields.Keys.All(k => s1.Fields.ContainsKey(k) && FieldComparer.Equals(s2.Fields[k], s1.Fields[k])))
-                {
-                    return false;
-                }
-
-                if (s1.HasMetadata && s2.HasMetadata)
-                {
-                    return s1.Metadata.Keys.Count() == s2.Metadata.Keys.Count() &&
-                           s1.Metadata.Keys.All(k => s2.Metadata.ContainsKey(k) && s1.Metadata[k] == s2.Metadata[k]) &&
-                           s2.Metadata.Keys.All(k => s1.Metadata.ContainsKey(k) && s2.Metadata[k] == s1.Metadata[k]);
-                }
-                return true;
-            }
-        }
         public class Build
         {
             [Fact]
